Battle UI Changes
For the Battle UI, I just made things a little bigger, as I was told that the buttons and robots are a bit small, I also added team names on top of the robots and repositioned the round sign to the top center of the screen. I also changed the watch sprites from the old one to new one.